change from delete hard to delete soft

+++ module/student/biz/delete_student.go
@@ -2,9 +2,15 @@ package studentbiz
 
 import (
 	"context"
+	"errors"
+	studentmodel "first-apis/module/student/model"
 )
 
 type DeleteStudentStore interface{
+	FindDataWithCondition(ctx context.Context, 
+		condition map[string]interface{},
+		moreKeys ...string,
+		) (*studentmodel.Student, error)
 	Delete(ctx context.Context, id int) error
 }
 
@@ -18,6 +24,15 @@ func NewDeleteStudentBiz(store DeleteStudentStore) *deleteStudentBiz {
 
 
 func (b *deleteStudentBiz) DeleteStudent(ctx context.Context, id int) error {
+	oldStudent, err := b.store.FindDataWithCondition(ctx,map[string]interface{}{"id": id})
+	if err!= nil {
+        return err
+    }
+
+	if oldStudent.Status == 0{
+		return errors.New("student does not exist")
+	}
+
 	if err := b.store.Delete(ctx, id); err != nil {
 		return err
 	}
